# Conflicts:
#	.pre-commit-config.yaml
#	Dockerfile
#	apps/backend/api/paychecks/post.py
#	config/nginx/nginx.conf
#	docker-compose.yml
#	pyproject.toml
#	requirements.txt
#	shared/models/paycheck.py

+++ shared/models/paycheck.py
@@ -21,7 +21,7 @@ class PaycheckCategory(Enum):
 
 class PaycheckCreate(SQLModel):
     name: str | None = None
-    price: int | None = None
+    price: float | None = None
     pay_date: float | None = datetime.datetime.now(datetime.UTC).timestamp()
     category: PaycheckCategory | None = None
